move useDevtools to the atoms directory
